be/requirements.txt
django>=5.2
djangorestframework>=3.14
pm4py
pandas
rustxes